// WorkpackMVC/src/app/models/user.ts
export class User {
  UserID: number;
  Password: string;
  CompanyId: number;
  FirstName: string;
  LastName: string;
  EmailID: string;
  DepartmentID: number;
  DepartmentName: string;
  DesignationID: number;
  DesignationName: string;
  ReportingTo: number;
  isActive: boolean;
  profilePic: string;
}